fix casting function

+++ core/WordpressFetcher.php
@@ -42,7 +42,7 @@ class WordpressFetcher
 
     private function try_cast($value)
     {
-        if (is_numeric($value) && intval($value) == intval(floatval($value)))
+        if (is_numeric($value) && floatval($value) == floatval(intval($value)))
             return intval($value);
 
         if (is_numeric($value))
